requirements/dev.txt
asgiref==3.7.1
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
click==8.1.3
decorator==5.1.1
Django==4.2.1
django-filter==23.2
django-model-utils==4.3.1
djangorestframework==3.14.0
executing==1.2.0
geographiclib==2.0
geopy==2.3.0
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg==3.1.9
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.2
wcwidth==0.2.6
